requirements.txt
Django==2.1.2
psycopg2-binary==2.7.6
djangorestframework==3.8.2
djangorestframework-recursive==0.1.2